extract marker generation method
